junit common.io 0.0.1

+++ sinnori/core_build/src/main/java/kr/pe/sinnori/common/updownfile/FileBlockInformation.java
@@ -23,7 +23,7 @@ public class FileBlockInformation {
 			long startFileBlockNo,
 			long endFileBlockNo,
 			
-			boolean append, int fileBlockSize, int fileBlockNo) {
+			boolean append, int fileBlockSize, long fileBlockNo) {
 		if (fileBlockNo < 0) {
 			String errorMessage = String.format("sourceFileID[%d]::parameter fileBlockNo[%d] is less than zero",
 					sourceFileID, fileBlockNo);
